+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Closed #938: `page_sidebar()` and `page_navbar()` now use `fillable=False` by default. (#990)
 
+### New features
+
+* Added `@render.download` as a replacement for `@session.download`, which is now deprecated. (#977)
+
 ### Bug fixes
 
 * CLI command `shiny create`... (#965)